refactor(eventtap): split the evdev capture file by concern
internal/adapter/eventtap/linux/evdev_cgo.go had grown to 1,487 lines
holding six distinct concerns, including a uinput scroll *output*
device buried in the middle of an input-capture file — the single most
confusing placement in the Linux tree.

The file now splits along those concerns, pure moves only:

- evdev_cgo.go           core types, constants, device identification
- evdev_capture_cgo.go   device scanning, readers, grab lifecycle
- evdev_hotplug_cgo.go   inotify watcher folding new keyboards into
                         an active grab
- evdev_xkb_cgo.go       key-code to key/modifier name translation
- evdev_scroll_cgo.go    the uinput scroll output device, now clearly
                         labeled as output rather than capture
- evdev_session_cgo.go   the per-mode tap session: dispatch, chord
                         passthrough, shutdown

Each file carries the minimal cgo preamble it needs; no function
bodies changed.
